added replication test case

+++ mobile/ios/example/test17_replication.js
@@ -5,18 +5,20 @@ var _ = require('underscore'),
 
 exports.run_tests = function() {
     var db = server.databaseNamed('test17');
-    db.create();
+    db.deleteDatabase();
+    db.ensureCreated();
+
+    var replicationComplete = false, checkCount = 15;
 
     try {
       // start with one replication at the beginning
-      var repl = db.pullFromDatabaseAtURL('http://localhost:5984/books');
+      var repl = db.pullFromDatabaseAtURL('http://touchbooks.iriscouch.com/books');
       repl.addEventListener('progress', function(e) {
         Ti.API.info(JSON.stringify(e));
-        // Ti.API.info(String.format("replication progress: %d; status: %s; completed: %d; total: %d; error: %s", e.running, e.status, e.completed, e.total, JSON.stringify(e.error)));
       });
       repl.addEventListener('stopped', function(e) {
         Ti.API.info(JSON.stringify(e));
-        // Ti.API.info(String.format("replication stopped: %d; status: %s; completed: %d; total: %d; error: %s", e.running, e.status, e.completed, e.total, JSON.stringify(e.error)));
+        replicationComplete = true;
       });
       repl.start();
     }
@@ -25,10 +27,15 @@ exports.run_tests = function() {
         throw e;
     }
 
-    /*
-    setTimeout(function() {
-      db.deleteDatabase();
-      Ti.API.info("deleted db");
-    }, 60000);
-    */
+    var interval = setInterval(function() {
+      if (replicationComplete) {
+        Ti.API.info("replication done!");
+        clearInterval(interval);
+        db.deleteDatabase();
+      }
+      else if (checkCount-- < 0) {
+        throw Error("timed out waiting for replication");
+      }
+    }, 2000);
+
 }
